Move IndentSearch to a proper file
