+++ lib/dead_end/api.rb
@@ -201,6 +201,8 @@ require_relative "display_invalid_blocks"
 require_relative "balance_heuristic_expand"
 require_relative "parse_blocks_from_indent_line"
 
-require_relative "block_document"
 require_relative "block_node"
 require_relative "indent_tree"
+require_relative "block_document"
+
+require_relative "indent_search"
